Android: Add Workaround for camera on emulator
An application that uses a camera get stuck on an emulator, while
running properly on a real device. To make it work on emulator the
workaround was prepared. It will make a copy of each frame buffer,
so it is not recommended to by used by default.

Fixes: QTBUG-122118
Fixes: QTBUG-119472
Pick-to: 6.5
Change-Id: Ie03fd527146b3e5a47da3338a4fc5d0b5ae21bf5
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit c41a483e6383d79ae06927e2dde55f05665e7601)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>