Deprecate internal methods of QVideoFrame
We're going to expose QAbstractVideoBuffer, and the internal methods
of QVideoFrame doesn't work well with it.
- QVideoFrame::videoBuffer() is not needed as we don't want to
  get users access to it.
- The constructor (QAbstractVideoBuffer*, QVideoFrameFormat)
  is not needed as we'll add a new one with unique_ptr instead.

Let's use private API instead of the internal methods.
The private construction also allows to avoid dynamic cast upon
the construction.
After adding public API, some of changes, especially in tests,
might be reverted to use the new QVideoFrame's constructor.

Change-Id: Id25d9e7faf9e88ce7f57492ba173dd168d6cf1de
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>