Add QtVideo::MapMode, deprecate QVideoFrame::MapMode
QAbstractVideoBuffer is going to be exposed in the public API,
and it will use MapMode, so the enum should be in a common namespace.

The only drawback of the deprecation is that QVideoFrame::mapMode()
still returns QVideoFrame::MapMode. However, calling the method
will not cause deprecation warnings, because enum values are deprecated,
not the enum as is.

[ChangeLog] Added the namespace QtVideo::MapMode, deprecated QVideoFrame::MapMode.

Change-Id: I5248c76c01f07a33f0f5a67e7db2d5072af5383d
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>