Expose QAbstractVideoBuffer in public API
QAbstractVideoBuffer is needed for users to be able to provide
custom data to QVideoFrame. We already have QAbstractVideoBuffer as
a private implementation, and QAbstractVideoBuffer was public in Qt5.

Currently, the video buffer doesn't provide any functionality to
work with hw textures directly. If a user works with GPU textures,
they will need to map/unmap it by overriding matching methods.

[ChangeLog] Added QAbstractVideoBuffer that allows to provide custom data to QVideoFrame.
Added a QVideoFrame's constructor taking QAbstractVideoBuffer.

Change-Id: Iba5658f228a2373124c571ff8cf71595ef970b43
Reviewed-by: Tim Blechmann <tim@klingt.org>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>