Tests: MediaFileSelector - Accept `BufferingMedia` as "finished" state
We don't need to fill the full buffer. Most notably this seems to cause
suprious unit test failures with the gstreamer backend on CI. However
when the media player is `BufferingMedia` we know that the media is
supported.

Pick-to: 6.5
Change-Id: I8880d5ba3121e2e8154e5c15e3cdc4fbaa4b5014
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
(cherry picked from commit 03d9885bfda6e2eeeb848e98cdf59559c79aaf48)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>