Fix wrong frame rate and frame count when encoding custom video frames
This patch sets duration of AVPackets before they are sent from video
encoding to muxer. This fixes issues where single-frame videos could not
be played in ffplay and QMediaPlayer. It also fixes issues with lost
frames and incorrect media duration.

It is not clear why FFmpeg muxer is unable to produce valid video files
without duration estimates, but internal muxer function
guess_pkt_duration appears to be assuming that either duration is
already set, or avg_frame_rate is set on the stream before
avformat_write_header is called.

Fixes: QTBUG-125957
Pick-to: 6.8
Change-Id: Ibb6853a356242271be404f26715095be835a1924
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>