Call QAbstractVideoBuffer::unmap in the destructor of QVideoFramePrivate
Unmap is a virtual function and it mustn't be called in destructors.
Before this patch, many of buffer destructors called the exact
overload CurrentBuffer::unmap in the destructor, which is safe, but
a bit awkward. The patch makes the destructor of the buffer holder
responsible for calling 'unmap' if the buffer is still mapped.

The patch will allow to handle user-made buffers properly.

The patch removes awkward calls of virtual unmap from

Change-Id: I69880e077c5e365b7a84f73c8e1349cffb7695c7
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit efb03f5993c8b756b86fe7e1456b34f9407401ff)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>