Android: Get the entirety of the UV-plane buffer of camera frames
Android Image objects have planes in order of: Y, U, V. We pointed the
UV-plane to the beginning of the U-plane buffer, which in the case of
NV21 is after the beginning of the V-plane buffer, thus causing
artifacts at the ends of video frame lines.

This patch adds detection between NV12 and NV21, and ensures that the
entirety of the UV-plane is included.

Fixes: QTBUG-123220
Pick-to: 6.7 6.5
Change-Id: I9771e0a5d80d8a113ce1fb72cc05fe12a8613c76
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 1cb9e7757751d74e7c25a12543dc1df02eeb80a4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>