Copy frame textures out from FFmpeg frame pool when decoding on Windows
The FFmpeg decoder frame pool is easily starved if video frames are kept
around for too long.

This patch fixes this issue on Windows by copying the frame textures
from the decoder frame pool as soon as we receive them from the decoder.
This way, the AVFrame can immediately be returned to the decoder's frame
pool. This can in some cases reduce GPU memory usage because we no
longer need to extend the size of the FFmpeg's decoder pool.

Task-number: QTBUG-125956
Change-Id: Iad6de0b932ada8b2e414ecdf4c7c410ef18c6f8f
Pick-to: 6.8
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>