Hide Mediaplayer PlaybackControl item automatically also when paused
The MediaPlayerExample is useful for comparing image quality with other
media players such as VLC. This is often done in full-screen pause mode
by alt-tabbing between VLC and media player. It is then beneficial that
the PlaybackControl item auto-hides in pause mode because this makes it
easier to compare image quality. The timeout for hiding the
PlaybackControl item is also reduced from 3 to 1.5 seconds because this
seems sufficient.

Pick-to: 6.8 6.7
Change-Id: Iafbe5fa31b2776b4c5038dafbef23294d8b247fc
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Ed Cooke <ed.cooke@qt.io>