AVFMediaPlayer: set the initial audio output device on new player
Each time we set a new source in AVFMediaPlayer, we recreate the
AVPlayer object and initialize it according to the media session
settings. Thus, we set audio output setting, such as volume or muted,
but for some reason we forget to initialize audio input device, which
leads to the issue that a media is always played through default audio
output device, instead of the device explicitly selected by an user.
So we will initialize it here accordingly.

Fixes: QTBUG-126203
Pick-to: 6.8 6.7 6.5
Change-Id: I309afec2de9aa00bb4df79eb2c90dd906d6ebc20
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>