GStreamer: audio io - make device switching more robust
Instead of adding a new src or sink elements and swapping the elements
in the pipeline, we set the "device" property on the currently existing
element when possible. This reduces the necessity to perform pipeline
changes and the potential for starvation.

Pick-to: 6.7 6.5
Change-Id: I6501bbd5151d5b26c1d39e23f27610fa4924d8d4
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 1692ccfb1123b4ec9fd2f35f41b83a2d7709d6ad)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>