GStreamer: rework stopped state handling
We set the gstreamer pipeline into "paused" state, while we still
consider it "stopped" (as in, we show a black frame). The original
approach of adding a flag to the GstPipeline was a bit problematic:
* it broke encapuslation
* it did not propagate the state change actively, but required a
pipeline flush.

We move the functionality out of the pipeline into the
`QGstVideoRenderer` and propagate the video output activation
explicitly.

Pick-to: 6.5 6.7 6.8
Change-Id: I8f847cd7296374cdf86e139dfe82c3142540bb37
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>