GStreamer: media player - rework seek-on-play logic
The QGstreamerMediaPlayer was flushing the pipeline during startup. This
caused the position to be queried from the pipeline, which is not
totally necessary (and caused problems in the past).
It's simpler and more robust to track the position that we want to start
at.

Pick-to: 6.7 6.5
Change-Id: I2186d46579ae022e484501adce1fea1ab7fe4534
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 925571fea0e351c7f4626cc9902796bd90f54af9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>