CMake: Rename QGstreamerMediaPluginPrivate module for SBOM purposes
For SBOM (Source Bill of Material) generation purposes, there is an
implicit assumption that Foo and FooPrivate target names map to the
same semantic type, e.g. Core and CorePrivate map to a module and the
module's private headers.

This breaks down if the types are different, which is the case
for QGstreamerMediaPluginPrivate which is a module and
QGstreamerMediaPlugin is a plugin.

Until the SBOM code is taught to be smarter, it's easiest to just
rename the targets. This should be fine, because the module is
internal.

Pick-to: 6.7 6.5
Change-Id: I9acd62d5bdead2e15c1a3cb13f5d82cf2efc0604
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit da0d6a04d7c91069fc2293942eaed6be195c2364)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>