wasm: add window and screen capture
This add support for window and screen capturing.

This is restricted for user with secure contexts, as it uses getDisplayMedia to
grab a window, and browser tab or an entire display screen, and requires
user permission to share that part of the screen.

This works differently than on other platforms, as the permissions cannot be
persisted, so every session requires the user to
give permission each time.
see
https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getDisplayMedia
and
https://developer.mozilla.org/en-US/docs/Web/API/Screen_Capture_API/Using_Screen_Capture

Also, there is no real way to restrict the capture to window, or screen
or application, the user can give permissions to one or more, or none.

Fixes: QTBUG-126602
Change-Id: Ie9ce99e8e3fc135f57d56f50edae95c50fa0307a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>