GStreamer: fix seeking logic
Gstreamer has undefined behavior when seeking on a pipeline that has
pending changes because (a) the pipeline is modified or (b) because the
of flushing seek operations.
We can guard against this class of bugs by trying to ensure that
get_state() does not return GST_STATE_CHANGE_ASYNC. Furthermore we
cannot set the playback rate as long as we cannot query
position/duration from the pipeline due to a GStreamer upstream bug.

Pick-to: 6.5 6.7 6.8
Change-Id: Ica2e2a9e33971a8c5d6f21051ea248996b5411c7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>