Override the signal 'videoFrameChanged' in qml video sink
Even though VideoSink is not documented in qml, some users utilise
the signal videoFrameChanged. The lifetime of the emitted frame is
managed by the qml garbage collector. In practice, it varies up to 3sec.
If the frame is gpu-based, keeping them alive for a long time can cause
media playback issues, especially with high (60) fps.
The patch overrides the signal in qml to get round the described issue.

[ChangeLog] Removed QVideoFrame parameter from the not documented qml signal VideoSink.videoFrameChanged.

Task-number: QTBUG-122757
Task-number: QTBUG-126106
Pick-to: 6.8 6.7 6.5
Change-Id: I7af977f28e9f1331bd000d245398483dee0c955f
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>