FFmpeg: Add default codecs to container codec lists
We call avformat_query_codec() to check for compatibility between
container and codec. If this returns a negative number, the test is
inconclusive, but in the case of default codecs, we can assume that they
are supported anyways.

Task-number: QTBUG-126143
Pick-to: 6.7 6.5
Change-Id: I99cd709a5016a450b0361faab122313bbbf03518
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit a84ce3af641c4cab713d4034888d8590ce554c27)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>