AVFMediaEncoder: Fix crash with when recording with stereo audio source
When creating NSData from an AudioChannelLayout structure, we specified
the length (number of bytes to copy) without accounting for the variable
length array mChannelDescription inside the AudioChannelLayout.
This caused a crash when recording with a camera with multiple audio
channels, because AVAssetWriterInput was getting uninitialised data when
reading the mChannelDescriptions array.

Fixes: QTBUG-126254
Pick-to: 6.8 6.7 6.5
Change-Id: Ia13dd3654250ba79aa6000624e84a4bebe40e850
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>