FFmpeg: Discard packets with timestamp after stream end point
Some files can be corrupted so that the packets decoded by FFmpeg end up
with malformed duration field. These packets are not considered corrupt
by FFmpeg and are not dropped and don't come with the
AV_PKT_FLAG_CORRUPT flag. This causes issues in our demuxing and
looping timings which are tied to the presentation timestamp and
duration of the packets.

However, these packets are observed to be outside of the duration value
of the stream they belong to, when this duration is provided in the
source file. This patch adds logic to consider any packets with a
presentation timestamp after the end of stream as EOF.

Fixes: QTBUG-126592
Pick-to: 6.8 6.7 6.5
Change-Id: I1236ac61fdd304d4c38d9a22e5f059119f9d54fc
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>