Revert "Add QtVideo::MapMode, deprecate QVideoFrame::MapMode"
This reverts commit 0fbdd3f5776d9603d817315b613b888d428a24c0. We
cannot deprecate an enum value that is used as a return type of
a getter, without also deprecating the getter. QtVideo::MapMode
was a scoped enum so didn't implicitly convert to int or
QVideoFrame::MapMode (and visa versa), and we cannot change the
return type of a function anyway.

Switch statements over the return value of the getter but using
the QtVideo::MapMode enum values (to avoid deprecation warnings)
would require static casts somewhere as well.

Given that QVideoFrame::MapMode is only used in APIs that are
tightly related to video frame management (QAbstractVideoBuffer
is used with a QVideoFrame), it's ok to leave the old enum as
is.

As a drive-by, use Q_ENUM to register the MapMode enum with the
meta type system, instead of implementing an equivalent QDebug
streaming operator.

Found during header review.

Pick-to: 6.8
Task-number: QTBUG-125859
Change-Id: Id94a3aa1e49405bca0b3f93b54302b57de8c6590
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>