FFmpeg decoder: Retry receiveAVFrames for EAGAIN if stream is end
[0]avcodec_receive_frame function may return EAGAIN error. It indicates
that the decoder needs more packets before it can produce a frame. Some
codecs, especially H.264, might need several packets to start producing
frames.

What is more, for FFAMediaCodec on Android, FFmpeg relies on
[1]dequeueOutputBuffer JNI call with timeout. It seems that in
[2]ff_AMediaCodec_dequeueOutputBuffer call, the timeout set to 1 second
may not be always enough.

When all packets were send, there will be a flush packet (that signals
End of the Stream). After that there will be no more packets.
In case when all packets were sent and avcodec_receive_frame call
returns EAGAIN error, it is possible that the reason of the error is
just a timeout on dequeueOutputBuffer JNI call.

That is why, in this case, we will give one more try to call
avcodec_receive_frame.

[0]https://ffmpeg.org/doxygen/7.0/group__lavc__decoding.html#ga11e6542c4e66d3028668788a1a74217c
[1]https://developer.android.com/reference/android/media/MediaCodec#dequeueOutputBuffer(android.media.MediaCodec.BufferInfo,%20long)
[2]https://ffmpeg.org/doxygen/7.0/mediacodecdec__common_8c_source.html#l00726

Pick-to: 6.8 6.7 6.5
Fixes: QTBUG-126428
Change-Id: I0fbb6a7eb6fb225303537402085dbdca825c524c
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>