QSampleCache: Fix reloading of erroneous sample
If a sample's state in the cache is QSample::Error, the load is invoked
again for the sample. However, since the thread was not started, the
state remained stuck in QSample::Loading forever.

This causes issues for the application, as it cannot observe the error
the second time the sample is used. When the error is reliably reported,
the application can handle the situation appropriately. Furthermore, in
the case of a network error, reloading the sample might succeed on the
second attempt.

Pick-to: 6.5 6.2
Change-Id: I6051fdc1ab65158378e1b22ad774c085ce4c1537
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
(cherry picked from commit 410bf91598d6c6964198bfa0e929d87bb6a14e97)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 5567355f93ed337135e88f8c55fc4b84bf56475f)