CMake: Don't record a ffmpeg package dependency for Qt static builds
In a static Qt build, if QFFmpegMediaPlugin links to the ffmpeg
targets, this will record a dependency on the ffmpeg package for the
plugin. Then when a user projects does find_package(Qt6Multimedia),
if the ffmpeg libs and includes are not present on the machine,
configure will error out saying that ffmpeg can't be found.

The Qt ffmpeg plugin expects to dlopen the libraries, so we don't
really need to propagate the link requirement to user projects.

Change the code to use the ffmpeg include paths and compile
definitions, instead of linking to the library targets, thus avoiding
propagating the link requirement to user projects.

Fixes: QTBUG-126817
Change-Id: I58b62e78d1655002e8b89275eaa2340bed9cedcd
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit 2bfd6c90e285371a3c5ba1434dca3449b7321bfb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>