Remove double mirroring of video frames
In a21f37193f2c105a174db7334cdf8c7a7199b3cb commit, information about
mirroring video frame was moved from internal "mirror" variable to
QVideoFormat. In current implementation mirror of video frame and
its format represent the same.
Because of that, mirroring has double impact when displaying frame.
One of it is removed in this commit. It is not a good place to do it, as
the texture can still be rotated later. The second one is done later
in QSGVideoNode::setTexturedRectGeometry (after rotation).

Additionally added information about frame mirroring to
QAndroidCamera::frameFormat(). This allows to consider mirroring frames
on Android while recording.

Currently double mirroring is removed only for Android.

Pick-to: 6.8 6.7 6.5
Fixes: QTBUG-127533
Change-Id: I1761ab38eca2c5758af4113926af2065be63b918
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>