Don't run GPU tests if RHI rendering is unsupported
tst_qvideoframecolormanagement fails with offscreen QPA platform
because some of the tests requires RHI based rendering. If RHI based
rendering was not present, a less accurate CPU fallback was used for
YUV to RGB conversion. The best option is to disable tests that require
RHI rendering if RHI based rendering is not supported. We still run the
tests that verify the CPU fallback.

Fixes: QTBUG-127745
Fixes: QTBUG-127744
Fixes: QTBUG-127746
Task-number: QTBUG-127784
Pick-to: 6.8 6.7 6.5
Change-Id: Id80c199653fb8fd882d99365d6ea1b7d63725fbe
Reviewed-by: Liang Qi <liang.qi@qt.io>