Remove empty/file scheme check from qMediaFromUserInput
QUrl::fromUserInput that is called inside qMediaFromUserInput function
is already doing everything required to resolve file paths without a
scheme presuming it's "file://". The checks that are removed in this
change is causing undesirable behavior on Windows when absolute paths
are used - drive letter is parsed as scheme therefore scheme is
considered not empty which is wrong.

Fixes: QTBUG-127320
Pick-to: 6.8 6.7 6.5
Change-Id: Ifd3f93413e38e04d08196d37e8de79aad8133c99
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>