FFmpegAudioRenderer: Additional check when AudioSink state changed
When QAudio::IdleState is send from QAudioSink, QFFmpegAudioRenderer is
proceeding with nextStep of rendering.
If audioOutput was changed and now we will try to proceed with nextStep
- at the end the audiosink will be destroyed. That will cause the crash
(when trying to get back to signal emiter that was destroyed).

That is why we should not allow to proceed with nextStep from QAudioSink
signal connection if the device was changed.

Pick-to: 6.8 6.7
Fixes: QTBUG-127681
Change-Id: Ic6ed43232bc446b03fc8af9f4f02145149c9e279
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>