Fix blocky UYVY and YUYV rendering with Qt Multimedia
UYVY and YUYV formats packs two pixels into a single vec4 value,
resulting in a texture that is half the width of the output image. This
means that one single input pixel contributes to two neighboring output
RGBA pixels. When populating output RGBA pixel, we must therefore check
if we are on an even or odd column before determining which input pixel
we sample from.

This patch also elliminates branching in the shader at the potential
cost of sampling each input pixel twice, still this should be faster. We
also avoid integer arithmetics that could reduce performance on some
GPUs.

Fixes: QTBUG-128246
Pick-to: 6.8 6.7 6.5
Change-Id: If7c955e7d5a7037db3e01afe10a8747f639f38c6
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>