Add manual test with 'minimal' audio recording implementation
Having a minimal audio recorder can be instructive and a good starting
point for experimenting with the qt media capture session and audio
devices.

The example could be smaller if we don't add support for selecting
device, but having this feature is useful for debugging issues that are
only seen on specific devices.

Task-number: QTBUG-128336
Pick-to: 6.7 6.5
Change-Id: I9f04f810b59cb03466dceacc573561c31a641cce
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 97099557e9b63feef1d6439c8190905caf630e22)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>