Don't add file extension when the filename already has an extension
QMediaRecorder and QImageCapture previously added an extra file
extension to file names when the file extension did not match any of the
extensions for the MIME type. This created file names on the form
"filename.mp3.mpga" on some platforms, and "filename.mp3" on other
platforms, and did not allow the application to decide the file
extension. Since different MIME databases can have have different
extensions for the same MIME type, this led to confusing cross-platform
behavior.

With this patch, the application can choose not to provide an extension
to get one generated, but if it provides an extension, this is not
changed by Qt Multimedia.

[ChangeLog][QtMultimedia] QMediaRecorder::setOutputLocation and
QImageCapture::captureToFile will no longer alter the suffix/extension
if the provided path already includes an extension, even if this
extension does not align with the recommended extensions for the MIME
type.

Pick-to: 6.7 6.5
Fixes: QTBUG-129501
Change-Id: Ia485ec37d275d55110c58a01aa21986ff3860d02
Reviewed-by: Vladimir Belyavsky <belyavskyv@gmail.com>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit e829737e6cae3f541b81780fd2f0963548f5fb4d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>