Don't add file extension when the filename already has an extension
QMediaRecorder and QImageCapture previously added an extra file
extension to file names when the file extension did not match any of the
extensions for the MIME type. This created file names on the form
"filename.mp3.mpga" on some platforms, and "filename.mp3" on other
platforms, and did not allow the application to decide the file
extension. Since different MIME databases can have have different
extensions for the same MIME type, this led to confusing cross-platform
behavior.

With this patch, the application can choose not to provide an extension
to get one generated, but if it provides an extension, this is not
changed by Qt Multimedia.

This cherry-pick to 6.5 required the following manual conflict
resolution steps:
1. Manual conflict resolution in tests/auto/integration/CMakeLists.txt,
   to ensure that the new qmediarecorderbackend tests were included.
   Changes to tests in tests/auto/integration/qmediaframeinputsbackend
   were not picked because they require APIs that are not available on
   Qt 6.5
2. Deleting incoming files that belong to qmediaframeinputsbackend
   tests. They require APIs that are not available on Qt 6.5
3. Deleted new tst_qmediarecorderbackend.cpp unit tests because
   they depended on new functionality not present in Qt 6.5.

[ChangeLog][QtMultimedia] QMediaRecorder::setOutputLocation and
QImageCapture::captureToFile will no longer alter the suffix/extension
if the provided path already includes an extension, even if this
extension does not align with the recommended extensions for the MIME
type.

Fixes: QTBUG-129501
Change-Id: Ia485ec37d275d55110c58a01aa21986ff3860d02
Reviewed-by: Vladimir Belyavsky <belyavskyv@gmail.com>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit e829737e6cae3f541b81780fd2f0963548f5fb4d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 067dcda60620a05714c4d1447ef5c430eb273e58)