GStreamer: Media player - improve setVideoSink while playing
When setting a new video sink, the new sink won't see a new frame
immediately, but only after the pipeline advances. When the pipeline is
playing, the new frame will arrive very quickly, so it is sufficient to
only flush the pipeline when the switching the video sink while paused.

Pick-to: 6.5 6.7 6.8
Change-Id: Ic3ccf9d392fc48ed5014a77ff345be9dd7998a58
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>