[doc] Improve Qt Multimedia landing page
This patch aims at improving the Qt Multimedia Landing page and make it
easier to read and use.

Since the Qt FFmpeg media backend is now the default on most platforms,
most content about this backend is moved first. The role of the native
backends is toned down, and moved after the notes on the FFmpeg backend.

The documentation is also updated to include new APIs. The section on
features removed from the FFmpeg libraries is also clarified.

Pick-to: 6.8
Change-Id: I8e0034dddc97c8bc2ba0a1b2e3abdfe9b9b6924d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>