AVFVideoSink: Avoid race condition when accessing Metal texture cache
The metal texture cache is owned by AVFVideoSink, i.e. it's created
and destroyed on the main thread. It's also passed to AVFVideoBuffer
for use in texture creation, which happens on render thread.
This leads to the standard multi-threaded problem of accessing a
shared resource.

In f4afbf6c573f8d863f456f25b6cc9ac37e4de5c7 we tried to fix this
by adding a mutex. But this did not solve the problem in general,
because now the mutex object itself became a shared resource - it
belonged to AVFVideoSink and was used in AVFVideoBuffer at the same
time. Therefore, when locking it on render thread, we can't be sure
that mutex object won't be destroyed simultaneously in the main thread.

Then, in cb8f812e6cc6cab374075fdd81b5d6f5f40e8a40 we tried to fix
the remaining issue. But again, it didn't solve initial data race
problem, so the crashes continued.

With this change, we fix the original problem entirely, by avoiding
the data race in principle. In AVFVideoBuffer, instead of accessing
the texture cache from AVFVideoSink every time a texture is created,
we will store a reference to it directly in the constructor on the
same main thread.

Fixes: QTBUG-128374
Task-number: QTBUG-110131
Task-number: QTBUG-120465
Change-Id: I573587e2abc74034763a496529861c0dfc2df0a8
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit aadb05820e9b3ed12749520b1c0df96b51b02014)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d7329f8748fab8589ceaf6c1f0517ed3b018e48b)