Fix Q_ASSERT crash in QFFmpegMediaPlayer::setMediaAsync()
If QMediaPlayer::setSource() was called with an empty URL while media
was still being loaded asynchronously from the previous call to
setSource(), mediaStatus() returned NoMedia when the asynchronous
function returned, triggering the Q_ASSERT in debug builds.

With this change the cancelToken will be checked first, causing
the function to return before the Q_ASSERT.

Also added a test that confirms that the Q_ASSERT is not triggered
in the above scenario and that the cancelToken system is working.

Fixes: QTBUG-128159
Pick-to: 6.8 6.7 6.5
Change-Id: Ic0bef26ffc7b0f37302b5e337800de08dc38f3db
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>