Fix error message bug in Media Player example
In the current implementation of the Media Player example, the error
message that is shown when an error occurs appears in a MessageDialog.
The Dialog is opened before its message is updated. The intended use of
a Dialog is that its content can not be updated while it is open. This
patch ensures the message of the Dialog is updated before it is opened.

Fixes: QTBUG-128335
Pick-to: 6.7
Change-Id: I19b931adcf27a0cb2e99d876f7df19eacaaa97e6
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
(cherry picked from commit 14a479ae78fa08f7843d35ea94c4c82a3687a60f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>