Add tests for recording custom frames with presentation transformations
With the programmatic IO API QVideoFrameInput, the transformations set
on the QVideoFrame are presentation parameters for display, and should
not apply to the recorded data.

This patch adds tests to verify that presentation parameters do not
change the data being recorded to file.

Also, we remove a dead test.

Task-number: QTBUG-128838
Pick-to: 6.8
Change-Id: Ifde51c575781f7bc707455c06ae2c476e4927469
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>