Implement extracting mirrored flag in FFmpeg media player
Video stream contains matrix transformation (side data) that can
be interpreted as rotation and mirroring (hflip).
FFmpeg API is able to extract rotation from the transform matrixes
via av_display_rotation_get, which doesn't work correctly
of mirroring is set on top of rotation.

It means that we should have our own function that extracts both,
rotation and mirroring. The CR implements the function and covers
it by unit and auto tests.

Pick-to: 6.8
Change-Id: I46ef597a11718cc664ab8025efaab7f2a99d4cd7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>