Increase timeout in tst_QScreenCaptureBackend for Linux
Screen capture tests are flaky, and it looks like the color of the
displayed video is wrong for some time after window is displayed. This
patch increases the timeout for waiting for window to settle on Linux,
and removes the timeout on other platforms as it should not be
necessary.

Pick-to: 6.7 6.5
Change-Id: I1d068f2db4489565e7c3e101dda73d4a9060daeb
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit 9f68280aee815e3f2df1a4f99af94c58e30d7dd8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>