Add REUSE.toml files and missing licenses
REUSE.toml files are read by reuse to complement or override the
copyright and licensing information found in file.

The use of REUSE.toml files was introduced in REUSE version 3.1.0a1.
This reuse version is compatible with reuse specification
version 3.2 [1].
With this commit's files,
* The SPDX document generated by reuse spdx conforms to SPDX 2.3,
* The reuse lint command reports that the Qt project is reuse compliant.

In order to be reuse compliant all the licenses present in the
LICENSES directory must be used. The license files only mentioned in qt_attribution.json
are moved next to the qt_attribution.json file and renamed with LICENSE. prefixe to
be excluded from reuse check.

[1]: https://reuse.software/spec-3.2/

[ChangeLog][Third-Party Code] Add path to license files within LicenseFiles field.
qtattributionscanner looks for the files associated with licenses listed in the
licenseId fiel within the LICENSES directory. To prevent reuse from erroring out, the
license files only mentioned in qt_attribution.json need to be moved out of the
LICENSES directory. For the qtattributionscanner to find those files, they need to be
listed in LicenseFiles field.

Task-number: QTBUG-124453
Task-number: QTBUG-125211
Pick-to: 6.8
Change-Id: I937cb86c9db5479544c4d8431af179584a9acf0a
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit 156db2b7e4607e572762f829a9641842fdf8c053)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>