Revert "Android: Get the entirety of the UV-plane buffer of camera frames"
This reverts commit 2a484b9a756c9933b8b7e1c4d0797b6a56e0a086.

Reason for revert: The change seems to be correct, but still - on some
devices it caused issue with recording. While NV12 is widely supported,
NV21 may not be supported by the HEVC encoder on some devices. In this
case, it is better to return to the "green line" artifact than not be
able to record videos.

Task-number: QTBUG-123220
Task-number: QTBUG-128421
Change-Id: I3b0a3fc074f3add7172d117e42e94497d7081bed
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>