Extend tests of QVideoFrameInput to cover more RGB and YUV formats
This test extends the existing color tests to verify that the
QMediaCaptureSession and QMediaRecorder handles colors correctly with
custom input frames. As a refrence image, we use the video
generator's output frame, converted with QVideoFrame::toImage, which is
verified through other tests.

On CI, we store the reference images along with the actual images from
the media file to COIN_CTEST_RESULTSDIR, which allow us to inspect any
differences from the CI logs.

Note that this test extends existing tests for more pixel formats, but
does not verify that the pipeline supports custom frames with different
color spaces or color ranges.

Task-number: QTBUG-126154
Pick-to: 6.8
Change-Id: Ib66dcb70cb9786f79c03e16a6ba4ed54481a3fc4
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>