GStreamer: fix spurious deadlocks in capture session
stopAndRemoveElements seems to be dangerous from an idle probe. We
rework the logic to:

* never stop/remove elements in the probes, but instead just unlink
other elements and stop from the application thread.
* avoid syncChildrenState, which seems to misbehave if the element is in
a state transition. however, since the pipeline is a live pipeline we
know that it should be PLAYING, so we set the playing state explicitly.
and sync with the state change from outside the idle probe

Fixes: QTBUG-129712
Pick-to: 6.5
Change-Id: I410934b245344ca1b26b99e17c7274008b77a21f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
(cherry picked from commit c8ba6b001f59915f7aa71d83e94e34613aec8960)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>